+++ src/main/java/com/programmers/smrtstore/domain/keep/application/KeepService.java
@@ -1,5 +1,9 @@
 package com.programmers.smrtstore.domain.keep.application;
 
+import com.programmers.smrtstore.core.properties.ErrorCode;
+import com.programmers.smrtstore.domain.keep.domain.entity.Keep;
+import com.programmers.smrtstore.domain.keep.exception.KeepException;
+import com.programmers.smrtstore.domain.keep.infrastructure.KeepJpaRepository;
 import com.programmers.smrtstore.domain.keep.presentation.dto.req.CreateKeepRequest;
 import com.programmers.smrtstore.domain.keep.presentation.dto.req.DeleteKeepRequest;
 import com.programmers.smrtstore.domain.keep.presentation.dto.req.FindKeepByCategoryRequest;
@@ -7,19 +11,42 @@ import com.programmers.smrtstore.domain.keep.presentation.dto.res.CreateKeepResp
 import com.programmers.smrtstore.domain.keep.presentation.dto.res.DeleteKeepResponse;
 import com.programmers.smrtstore.domain.keep.presentation.dto.res.KeepRankingResponse;
 import com.programmers.smrtstore.domain.keep.presentation.dto.res.KeepResponse;
-import org.springframework.data.domain.PageRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-public interface KeepService {
+@Service
+@RequiredArgsConstructor
+public class KeepService {
+    private final KeepJpaRepository keepRepository;
 
-    CreateKeepResponse createKeep(CreateKeepRequest request);
+    public CreateKeepResponse createKeep(CreateKeepRequest request) {
+        Keep keep = Keep.builder()
+                .userId(request.getUserId())
+                .productId(request.getProductId())
+                .build();
+        Keep saveKeepEntity = keepRepository.save(keep);
+        return CreateKeepResponse.of(saveKeepEntity);
+    }
 
-    List<KeepResponse> getAllKeepsByUserId(Long userId);
+    public List<KeepResponse> getAllKeepsByUserId(Long userId) {
+        List<Keep> keepListFindByUserId = keepRepository.findAllByUserId(userId);
+        return keepListFindByUserId.stream().map(KeepResponse::of).toList();
+    }
 
-    DeleteKeepResponse deleteKeep(DeleteKeepRequest request);
+    public DeleteKeepResponse deleteKeep(DeleteKeepRequest request) {
+        Long deleteId = request.getId();
+        Keep deleteKeep = keepRepository.findById(deleteId).orElseThrow(() -> new KeepException(ErrorCode.KEEP_NOT_FOUND_ERROR));
+        keepRepository.delete(deleteKeep);
+        return DeleteKeepResponse.from(deleteId);
+    }
 
-    List<KeepResponse> findKeepByUserAndCategory(FindKeepByCategoryRequest request);
+    public List<KeepResponse> findKeepByUserAndCategory(FindKeepByCategoryRequest request) {
+        return null;
+    }
 
-    List<KeepRankingResponse> getKeepRanking(PageRequest pageRequest);
+    public List<KeepRankingResponse> getKeepRanking(int limit) {
+        return keepRepository.findTopProductIdsWithCount(limit);
+    }
 }
